refactoring address processor
