+++ service-front/module/Application/src/Controller/CPFlowController.php
@@ -14,6 +14,7 @@ use Application\Forms\PassportDate;
 use Application\Forms\PassportNumber;
 use Application\Forms\Postcode;
 use Application\Forms\PostOfficePostcode;
+use Application\Helpers\AddressProcessorHelper;
 use Application\Helpers\FormProcessorHelper;
 use Laminas\Form\Annotation\AttributeBuilder;
 use Laminas\Http\Response;
@@ -28,6 +29,7 @@ class CPFlowController extends AbstractActionController
         private readonly OpgApiServiceInterface $opgApiService,
         private readonly FormProcessorHelper $formProcessorHelper,
         private readonly SiriusApiService $siriusApiService,
+        private readonly AddressProcessorHelper $addressProcessorHelper,
         private readonly array $config,
     ) {
     }
@@ -378,7 +380,14 @@ class CPFlowController extends AbstractActionController
                     $params->get('postcode'),
                     $this->getRequest()
                 );
-                $addressStrings = $this->formProcessorHelper->stringifyAddresses($response);
+                $processedAddresses = [];
+                foreach ($response as $foundAddress) {
+                    $processedAddresses[] = $this->addressProcessorHelper->processAddress(
+                        $foundAddress,
+                        'siriusAddressType'
+                    );
+                }
+                $addressStrings = $this->addressProcessorHelper->stringifyAddresses($processedAddresses);
                 $view->setVariable('addresses', $addressStrings);
                 $view->setVariable('addresses_count', count($addressStrings));
                 return $view->setTemplate('application/pages/cp/select_address');
