+++ service-front/module/Application/test/Controller/IndexControllerTest.php
@@ -217,78 +217,4 @@ class IndexControllerTest extends AbstractHttpControllerTestCase
             $this->getResponse()->getBody()
         );
     }
-
-    /**
-     * @dataProvider addressData
-     */
-    public function testProcessAddress(
-        array $address,
-        array $addressType,
-        array $expected
-    ): void {
-        $siriusApiService = $this->createMock(SiriusApiService::class);
-        $opgApiService = $this->createMock(OpgApiService::class);
-
-        $indexController = new IndexController($opgApiService, $siriusApiService);
-
-        $response = $indexController->processAddress($address, $addressType);
-
-        $this->assertEquals($expected, $response);
-    }
-
-    public static function addressData(): array
-    {
-        return [
-            [
-                [
-                    'addressLine1' => "Park House",
-                    'addressLine2' => 'Park Lane',
-                    'town' => 'London',
-                    'postcode' => 'SW1A 1AA',
-                    'country' => 'UK'
-                ],
-                [
-                    'line1' => 'addressLine1',
-                    'line2' => 'addressLine2',
-                    'line3' => 'addressLine3',
-                    'town' => 'town',
-                    'postcode' => 'postcode',
-                    'country' => 'country'
-                ],
-                [
-                    'line1' => "Park House",
-                    'line2' => 'Park Lane',
-                    'line3' => '',
-                    'town' => 'London',
-                    'postcode' => 'SW1A 1AA',
-                    'country' => 'UK'
-                ]
-            ],
-            [
-                [
-                    'line1' => "Park House",
-                    'line3' => 'Park Lane',
-                    'town' => 'London',
-                    'postcode' => 'SW1A 1AA',
-                    'country' => 'UK'
-                ],
-                [
-                    'line1' => 'line1',
-                    'line2' => 'line2',
-                    'line3' => 'line3',
-                    'town' => 'town',
-                    'postcode' => 'postcode',
-                    'country' => 'country'
-                ],
-                [
-                    'line1' => "Park House",
-                    'line2' => '',
-                    'line3' => 'Park Lane',
-                    'town' => 'London',
-                    'postcode' => 'SW1A 1AA',
-                    'country' => 'UK'
-                ]
-            ],
-        ];
-    }
 }
