+++ service-front/module/Application/view/application/pages/cp/enter_address_manual.twig
@@ -1,39 +1,39 @@
 {% include 'layout/id_check_banner.twig' with details_data %}
 <h1 class="govuk-heading-xl govuk-!-width-two-thirds">What is the address on the ID document?</h1>
 <form method="POST" action="./confirm-address">
-    <div class="govuk-form-group {% if form.get('addressLine1').messages %}govuk-form-group--error{% endif %}">
-        {% if form.get('addressLine1').messages %}
-            <p id="addressLine1-error" class="govuk-error-message">
+    <div class="govuk-form-group {% if form.get('line1').messages %}govuk-form-group--error{% endif %}">
+        {% if form.get('line1').messages %}
+            <p id="line1-error" class="govuk-error-message">
                 <span class="govuk-visually-hidden">Error:</span>
-                {{ form.get('addressLine1').messages | join(', ', ' and ') }}
+                {{ form.get('line1').messages | join(', ', ' and ') }}
             </p>
         {% endif %}
-        <label for="addressLine1" class="govuk-label">Address line 1</label>
+        <label for="line1" class="govuk-label">Address line 1</label>
         <input
-                class="govuk-input govuk-!-width-one-third {% if form.get('addressLine1').messages %}govuk-input--error{% endif %}"
-                id="addressLine1" name="addressLine1" type="text"
-                value="{{ form.get('addressLine1').value }}"
-                aria-describedby="addressLine1-hint {% if form.get('addressLine1').messages %}addressLine1-error{% endif %}"
+                class="govuk-input govuk-!-width-one-third {% if form.get('line1').messages %}govuk-input--error{% endif %}"
+                id="line1" name="line1" type="text"
+                value="{{ form.get('line1').value }}"
+                aria-describedby="line1-hint {% if form.get('line1').messages %}line1-error{% endif %}"
         />
     </div>
 
     <div class="govuk-form-group">
-        <label for="addressLine2" class="govuk-label">Address line 2 (optional)</label>
+        <label for="line2" class="govuk-label">Address line 2 (optional)</label>
         <input
                 class="govuk-input govuk-!-width-one-third"
-                id="addressLine2" name="addressLine2" type="text"
-                value="{{ form.get('addressLine2').value }}"
-                aria-describedby="addressLine2-hint"
+                id="line2" name="line2" type="text"
+                value="{{ form.get('line2').value }}"
+                aria-describedby="line2-hint"
         />
     </div>
 
     <div class="govuk-form-group">
-        <label for="addressLine3" class="govuk-label">Address line 3 (optional)</label>
+        <label for="line3" class="govuk-label">Address line 3 (optional)</label>
         <input
                 class="govuk-input govuk-!-width-one-third"
-                id="addressLine3" name="addressLine3" type="text"
-                value="{{ form.get('addressLine3').value }}"
-                aria-describedby="addressLine3-hint"
+                id="line3" name="line3" type="text"
+                value="{{ form.get('line3').value }}"
+                aria-describedby="line3-hint"
         />
     </div>
